+++ Rentit-frontend/src/components/items/items.jsx
@@ -1,7 +1,7 @@
 import './items.css';
-import { useNavigate } from 'react-router-dom';
 import { useEffect, useState } from 'react';
 import Loading from '../../images/loading.svg';
+import ItemCard from './itemCard';
 // import { useGlobalContext } from '../../authContext';
 
 const Items = () => {
@@ -29,11 +29,6 @@ const Items = () => {
     fetchData();
   }, []);
 
-  const navigate = useNavigate();
-  const handleClick = (id) => {
-    user ? navigate('/itemDetail', { state: { id: id } }) : navigate('/user');
-  };
-
   if (!itemData)
     return (
       <>
@@ -46,25 +41,8 @@ const Items = () => {
     <>
       <div className='allItems'>
         {itemData &&
-          itemData.map((data) => {
-            return (
-              <div
-                className='item'
-                onClick={() => handleClick(data._id)}
-                key={data.id}
-              >
-                <div id='itemImgBox'>
-                  <img className='imgItem' src={data.image} alt='cover' />
-                </div>
-                <div className='line lineItems'></div>
-                <div className='itemInfo'>
-                  <div className='itemName'>
-                    <span>{data.product}</span>
-                  </div>
-                  <div className='itemDetail'>{data.description}</div>
-                </div>
-              </div>
-            );
+          itemData.map((data, index) => {
+            return <ItemCard key={index} data={data} user={user} />;
           })}
       </div>
     </>
